fix(codex): resolve lint failures in bridge-timing instrumentation
- Move CodexDynamicToolBridgeTiming's definition from protocol.ts to
  dynamic-tool-execution.ts (where the timing-attachment logic already
  lives) and reference it from protocol.ts via an inline type import.
  protocol.ts had crept to 706 code lines against its 700-line max-lines
  cap; this keeps the type in a natural home instead of trimming
  unrelated code, and lands protocol.ts at exactly 700.
- Split the JSON.parse(JSON.stringify(...)) round-trip assertion in
  dynamic-tool-execution.test.ts into two statements to satisfy
  unicorn/prefer-structured-clone without switching to structuredClone,
  which silently drops non-enumerable properties through a different
  mechanism than the one this test verifies (the actual JSON wire
  serialization path used by client.ts).

Local verification: pnpm lint:extensions clean, pnpm tsgo:extensions
clean, and all three affected Vitest configs still pass in full
(616 + 466 + 196 = 1278 tests).
